+++ docs/Whats-New.md
@@ -240,6 +240,7 @@ Vanilla fixes:
 - `[CombatDamage]->MinDamage` now works as expected and damage is no long always a minimum of `1` (by ZivDero)
 - VehicleTypes with Jumpjet locomotion now take damage in flight (by ZivDero)
 - Fix the map glitching around when scrolling if the map is not large enough to fill the entire screen (by Belonit, ZivDero)
+- Animations now use their `Warhead` to deal damage, if one is specified (by ZivDero)
 
 </details>